refactor: service에 hilt 적용
