--- Doraboda/app/src/main/java/com/univ/doraboda/Application.kt
@@ -1,11 +1,0 @@
-package com.univ.doraboda
-
-import android.app.Application
-import timber.log.Timber
-
-class Application : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        Timber.plant(Timber.DebugTree())
-    }
-}
