+++ Doraboda/app/src/main/java/com/univ/doraboda/DorabodaApplication.kt
@@ -1,0 +1,13 @@
+package com.univ.doraboda
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class DorabodaApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        Timber.plant(Timber.DebugTree())
+    }
+}
