+++ Doraboda/build.gradle.kts
@@ -3,4 +3,5 @@ plugins {
     id("com.android.application") version "8.13.0" apply false
     id("org.jetbrains.kotlin.android") version "2.1.10" apply false
     id("com.google.devtools.ksp") version "2.1.10-1.0.31" apply false
+    id("com.google.dagger.hilt.android") version "2.57.1" apply false
 }
